refactor(client): colocate migrated submission mutations as typed documents
Move the 13 submission/comment mutations off the central untyped `gql`
registry in mutations.ts into typed `graphql()` documents colocated in
their sole consuming components:

- SubmissionTitle.vue        UpdateSubmissionTitle
- ConfirmStatusChangeDialog  UpdateSubmissionStatus
- AssignedSubmissionUsers    reviewers/coordinators/submitters (+ relatedUserFields fragment)
- forms/CommentEditor        the 8 inline/overall comment ops (+ commentFields fragment)

The shared relatedUserFields/commentFields fragments are colocated too and
resolve across components via the client-preset fragment registry. Consumers
now use the generated `*Document` types, so mutation variables and results
are type-checked (surfaced and fixed a string-vs-SubmissionStatus gap).

vue-tsc, eslint, and the full vitest suite (506) pass.
