+++ client/src/components/AssignedSubmissionUsers.vue
@@ -66,6 +66,85 @@
   </section>
 </template>
 
+<script lang="ts">
+import { graphql } from "src/graphql/generated"
+
+// Colocated here as the natural home for the shared related-user shape;
+// other colocated operations (e.g. commentFields) reference it by name via
+// the client-preset fragment registry.
+graphql(`
+  fragment relatedUserFields on User {
+    id
+    display_label
+    username
+    name
+    email
+    avatar_color
+    staged
+  }
+`)
+
+graphql(`
+  mutation UpdateSubmissionReviewers(
+    $id: ID!
+    $connect: [ID!]
+    $disconnect: [ID!]
+  ) {
+    updateSubmissionReviewers(
+      input: {
+        id: $id
+        reviewers: { connect: $connect, disconnect: $disconnect }
+      }
+    ) {
+      id
+      reviewers {
+        ...relatedUserFields
+      }
+    }
+  }
+`)
+
+graphql(`
+  mutation UpdateSubmissionReviewCoordinators(
+    $id: ID!
+    $connect: [ID!]
+    $disconnect: [ID!]
+  ) {
+    updateSubmissionReviewCoordinators(
+      input: {
+        id: $id
+        review_coordinators: { connect: $connect, disconnect: $disconnect }
+      }
+    ) {
+      id
+      review_coordinators {
+        ...relatedUserFields
+      }
+    }
+  }
+`)
+
+graphql(`
+  mutation UpdateSubmissionSubmitters(
+    $id: ID!
+    $connect: [ID!]
+    $disconnect: [ID!]
+  ) {
+    updateSubmissionSubmitters(
+      input: {
+        id: $id
+        submitters: { connect: $connect, disconnect: $disconnect }
+      }
+    ) {
+      id
+      submitters {
+        ...relatedUserFields
+      }
+    }
+  }
+`)
+</script>
+
 <script setup lang="ts">
 import ReinviteUserDialog from "./dialogs/ReinviteUserDialog.vue"
 import FindUserSelect from "./forms/FindUserSelect.vue"
@@ -74,12 +153,14 @@ import UserList from "./molecules/UserList.vue"
 import { useFeedbackMessages } from "src/use/guiElements"
 import { useMutation } from "@vue/apollo-composable"
 import {
-  UPDATE_SUBMISSION_REVIEWERS,
-  UPDATE_SUBMISSION_REVIEW_COORDINATORS,
-  UPDATE_SUBMISSION_SUBMITERS,
   INVITE_REVIEWER,
   INVITE_REVIEW_COORDINATOR
 } from "src/graphql/mutations"
+import {
+  UpdateSubmissionReviewersDocument,
+  UpdateSubmissionReviewCoordinatorsDocument,
+  UpdateSubmissionSubmittersDocument
+} from "src/graphql/generated/graphql"
 import { computed, ref } from "vue"
 import type { DocumentNode } from "graphql"
 import { useI18nPrefix } from "src/use/i18nPrefix"
@@ -128,16 +209,16 @@ interface SubmissionMutationVars {
 
 const mutations: Record<string, RoleMutations> = {
   reviewers: {
-    update: UPDATE_SUBMISSION_REVIEWERS,
+    update: UpdateSubmissionReviewersDocument,
     invite: INVITE_REVIEWER
   },
   review_coordinators: {
-    update: UPDATE_SUBMISSION_REVIEW_COORDINATORS,
+    update: UpdateSubmissionReviewCoordinatorsDocument,
     invite: INVITE_REVIEW_COORDINATOR
   },
   submitters: {
-    update: UPDATE_SUBMISSION_SUBMITERS,
-    invite: UPDATE_SUBMISSION_SUBMITERS // TODO: Enable submitter invitation
+    update: UpdateSubmissionSubmittersDocument,
+    invite: UpdateSubmissionSubmittersDocument // TODO: Enable submitter invitation
   }
 }
 const setMutationType = computed(() => {
